# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "AuctionAdminAssign": {
        "additionalProperties": false,
        "description": "Admin assigns a provider to a quotation (with a suggested price).\n\nCreates a new Auction with state=PENDING and the admin's price as\nthe initial bid. The provider can then accept as-is, counter-offer,\nor decline.",
        "properties": {
          "admin_budget": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "description": "Admin's suggested price (MXN)",
            "title": "Admin Budget"
          },
          "id_truck": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Truck"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "people": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "People"
          }
        },
        "required": [
          "admin_budget"
        ],
        "title": "AuctionAdminAssign",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Message": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "Message",
        "type": "object"
      },
...
      "QuotationRead": {
        "properties": {
          "channel_sales": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Sales"
          },
          "client_email": {
            "title": "Client Email",
            "type": "string"
          },
          "client_name": {
            "title": "Client Name",
            "type": "string"
          },
          "client_phone": {
            "title": "Client Phone",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "destination_adress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Adress"
          },
          "destination_floor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Floor"
          },
          "destination_postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Postal Code"
          },
          "destination_pulley": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Pulley"
          },
          "destination_restrictions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Restrictions"
          },
          "destination_transport_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Transport Type"
          },
          "destination_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Type"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "id_saler": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Saler"
          },
          "items": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "origin_adress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Adress"
          },
          "origin_floor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Floor"
          },
          "origin_postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Postal Code"
          },
          "origin_pulley": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Pulley"
          },
          "origin_restrictions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Restrictions"
          },
          "origin_transport_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Transport Type"
          },
          "origin_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Type"
          },
          "products": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Products"
          },
          "saler": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saler"
          },
          "service_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Date"
          },
          "service_hour": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Hour"
          },
          "service_internal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Internal"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "service_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Type"
          },
          "service_zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Zone"
          },
          "services": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Services"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "wizard_complete": {
            "default": false,
            "title": "Wizard Complete",
            "type": "boolean"
          },
          "wizard_step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wizard Step"
          }
        },
        "required": [
          "id",
          "client_name",
          "client_phone",
          "client_email",
          "created_at",
          "updated_at"
        ],
        "title": "QuotationRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FastAPI port of the Mobbit B2B/B2C marketplace Rust Lambdas. See `docs/research/business-domain.md` for the business context and `docs/research/deployment-drift.md` for the original infra.",
    "title": "Mobbit Backend Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/quotation": {
      "get": {
        "operationId": "list_quotations_api_admin_quotation_get",
        "parameters": [
          {
            "description": "Filter by state",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by state",
              "title": "State"
            }
          },
          {
            "description": "Search by client name, email, phone, service name, or postal code",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by client name, email, phone, service name, or postal code",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuotationRead"
                  },
                  "title": "Response List Quotations Api Admin Quotation Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List quotations (admin)",
        "tags": [
          "admin:quotations"
        ]
      },
      "post": {
        "operationId": "create_quotation_api_admin_quotation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationCreateAdmin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a quotation (admin)",
        "tags": [
          "admin:quotations"
        ]
      }
    },
    "/api/admin/quotation/{quotation_id}": {
      "delete": {
        "operationId": "delete_quotation_api_admin_quotation__quotation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a quotation (admin)",
        "tags": [
          "admin:quotations"
        ]
      },
      "get": {
        "operationId": "get_quotation_api_admin_quotation__quotation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get a single quotation",
        "tags": [
          "admin:quotations"
        ]
      },
      "put": {
        "operationId": "update_quotation_api_admin_quotation__quotation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update a quotation",
        "tags": [
          "admin:quotations"
        ]
      }
    },
    "/api/admin/quotation/{quotation_id}/assign-provider": {
      "post": {
        "description": "Creates a new Auction with state=PENDING. The `provider_id` comes in the body. The `admin_budget` is the price the admin sets. The provider can then accept, counter-offer, or decline.\n\nReturns 409 if the provider already has an auction for this quotation.",
        "operationId": "assign_provider_api_admin_quotation__quotation_id__assign_provider_post",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          },
          {
            "description": "The provider to assign",
            "in": "query",
            "name": "provider_id",
            "required": true,
            "schema": {
              "description": "The provider to assign",
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionAdminAssign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign a provider to a quotation (with admin budget)",
        "tags": [
          "admin:quotations"
        ]
      }
    },
    "/api/admin/quotation/{quotation_id}/cancel": {
      "post": {
        "operationId": "cancel_quotation_api_admin_quotation__quotation_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel a quotation",
        "tags": [
          "admin:quotations"
        ]
      }
    },
...
    "/api/admin/quotation/{quotation_id}/publish": {
      "post": {
        "operationId": "publish_quotation_api_admin_quotation__quotation_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish a quotation (DRAFT \u2192 QUOTED)",
        "tags": [
          "admin:quotations"
        ]
      }
    },
...
    "/api/admin/search/quotations": {
      "get": {
        "description": "Full-text search across quotations.",
        "operationId": "search_q_api_admin_search_quotations_get",
        "parameters": [
          {
            "description": "Search query (name, phone, email, address, ID)",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Search query (name, phone, email, address, ID)",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuotationRead"
                  },
                  "title": "Response Search Q Api Admin Search Quotations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Q",
        "tags": [
          "admin:search"
        ]
      }
    },
...
    "/api/b2c/quotation": {
      "post": {
        "description": "Public endpoint. The client only submits their contact info.\nThe full quotation (addresses, services, items) is filled in by\nthe admin via the RCCM API.",
        "operationId": "create_quotation_api_b2c_quotation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationCreateB2C"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a public quotation (B2C lead)",
        "tags": [
          "b2c:quotations"
        ]
      }
    },
    "/api/b2c/quotation/{quotation_id}": {
      "get": {
        "operationId": "get_quotation_api_b2c_quotation__quotation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single quotation",
        "tags": [
          "b2c:quotations"
        ]
      },
      "put": {
        "operationId": "update_quotation_api_b2c_quotation__quotation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "quotation_id",
            "required": true,
            "schema": {
              "title": "Quotation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a quotation (B2C)",
        "tags": [
          "b2c:quotations"
        ]
      }
    },
...
  }
}